split save_events into replace and append events so not to burden the user with the Optional range
